Fix build with gas.
Do not use C constant suffixes.  Bit values are small enough to not
require typing, despite they are used for 64bit MSR writes.  The added
cast in hw_ibrs_recalculate() is redundand but I prefer to add it for
clarity.

Reported by:	bde
Sponsored by:	The FreeBSD Foundation
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@329214 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/x86/x86/cpu_machdep.c
@@ -592,7 +592,7 @@ hw_ibrs_recalculate(void)
 	if ((cpu_ia32_arch_caps & IA32_ARCH_CAP_IBRS_ALL) != 0) {
 		if (hw_ibrs_disable) {
 			v= rdmsr(MSR_IA32_SPEC_CTRL);
-			v &= ~IA32_SPEC_CTRL_IBRS;
+			v &= ~(uint64_t)IA32_SPEC_CTRL_IBRS;
 			wrmsr(MSR_IA32_SPEC_CTRL, v);
 		} else {
 			v= rdmsr(MSR_IA32_SPEC_CTRL);
